Rename package `addtwonumbers` to `service`.
